projects/parkinglot/requirements.txt
# capture.py
picamera
azure.storage.blob

# post_id.py
requests
